requirements.txt
click==6.7
colorama==0.3.9
decorator==4.2.1
fpdf==1.7.2
networkx==2.1
Pillow==5.0.0
PyPDF2==1.26.0
six==1.11.0
zip==0.0.1
